+++ bin/wafr.js
@@ -2,13 +2,28 @@
 
 const meow = require('meow');
 const path = require('path');
+const fs = require('fs');
+const utils = require('../src/utils/index.js');
 const wafr = require('../src/index.js');
 
 // handle cli
 const cli = meow(`
     Usage
       $ wafr <path to contract test>
-`);
+
+    Options
+      --output, -o  solc compile output to JSON
+
+    Examples
+      $ wafr ./contracts --output ./build/contracts.json
+`, {
+  alias: {
+    o: 'output',
+  },
+});
+
+// output path
+const outputPath = path.resolve(cli.flags.output);
 
 // the main wafr code to run
 wafr({ path: path.resolve(cli.input[0]), optimize: 1 }, (wafrError, wafrResult) => {
@@ -22,7 +37,18 @@ wafr({ path: path.resolve(cli.input[0]), optimize: 1 }, (wafrError, wafrResult) 
     if (wafrResult.status === 'failure') {
       process.exit(1);
     } else {
-      process.exit(0);
+      // if there is an output path, build sources
+      if (typeof outputPath === 'string') {
+        fs.writeFile(outputPath, JSON.stringify(utils.filterTestsFromOutput(wafrResult.contracts)), (writeFileError) => {
+          if (writeFileError) {
+            throw new Error(`while writting output JSON file ${outputPath}: ${writeFileError}`);
+          }
+
+          process.exit(0);
+        });
+      } else {
+        process.exit(0);
+      }
     }
   }
 });