+++ src/tests/solidityTests/dirDepth/~somebadfile.sol
@@ -1,0 +1,1 @@
+jjust another bad file..